+++ src/components/Loader/Loader.scss
@@ -3,7 +3,7 @@
   justify-content: center;
   align-items: center;
   width: 100vw;
-  height: 60vh;
+  height: 50vh;
   // padding: 100px 0 100px 0;
 }
