+++ src/components/ProductItem/ProductItem.scss
@@ -30,6 +30,7 @@
 }
 
 .product__img-container {
+  height: 281px;
   display: flex;
   justify-content: center;
   align-items: center;
